requirements.txt
flask==3.0.3
apscheduler==3.10.4
requests==2.32.3
gunicorn==22.0.0
pytz==2024.1
Pillow==10.4.0
psycopg2-binary==2.9.9
